chore: add change status update and cancel funtionalities
Added quick edit undo button
Separated home page components to diff files
Added sample ticktets json file
