+++ src/components/TicketTable.tsx
@@ -1,0 +1,78 @@
+import { useTicketContext } from "@/contexts/TicketContext";
+import type { TicketType } from "@/types";
+import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from "@/components/ui/table";
+import { Checkbox } from "@/components/ui/checkbox";
+
+const TicketTable = () => {
+    const { displayTickets, selectedTicketIDs, setSelectedTicketIDs } = useTicketContext();
+
+    const toggleRowSelection = (id: string) => {
+        setSelectedTicketIDs((prev) => {
+            const selectedTicketIDsCopy = new Set(prev);
+            if (selectedTicketIDsCopy.has(id)) {
+                selectedTicketIDsCopy.delete(id);
+            } else {
+                selectedTicketIDsCopy.add(id);
+            }
+            return selectedTicketIDsCopy;
+        });
+    };
+
+    const chooseStatusColor = (status: TicketType["status"]) => {
+        switch (status) {
+            case "Open":
+                return "bg-gray-500";
+            case "In progress":
+                return "bg-blue-500";
+            case "Resolved":
+                return "bg-green-500";
+            case "Closed":
+                return "bg-yellow-500";
+            default:
+                return "";
+        }
+    };
+
+    return (
+        <div className="h-full flex flex-col gap-2">
+            <Table className="bg-card w-full text-sm lg:text-lg">
+                <TableHeader className="bg-muted h-15 sticky top-0 z-1">
+                    <TableRow key="header" className="text-lg lg:text-xl">
+                        <TableHead className="text-primary pl-5">Edit</TableHead>
+                        <TableHead className="text-primary">Ticket</TableHead>
+                        <TableHead className="text-primary">Title</TableHead>
+                        <TableHead className="text-primary">Status</TableHead>
+                        <TableHead className="text-primary">Submitted At</TableHead>
+                        <TableHead className="text-primary text-right pr-5">Assigned To</TableHead>
+                    </TableRow>
+                </TableHeader>
+
+                <TableBody>
+                    {displayTickets.map((ticket) => (
+                        <TableRow key={ticket.id} onClick={() => toggleRowSelection(ticket.id)}>
+                            <TableCell className="pl-5">
+                                <Checkbox checked={selectedTicketIDs.has(ticket.id)} className="size-5 bg-muted" />
+                            </TableCell>
+                            <TableCell>{ticket.id}</TableCell>
+                            <TableCell>{ticket.title}</TableCell>
+                            <TableCell>
+                                <span className={`${chooseStatusColor(ticket.status)} p-1 rounded-xl text-center`}>
+                                    {ticket.status}
+                                </span>
+                            </TableCell>
+                            <TableCell>{ticket.created_at}</TableCell>
+                            <TableCell className="text-right pr-5">{ticket.assigned_to}</TableCell>
+                        </TableRow>
+                    ))}
+                </TableBody>
+            </Table>
+            {displayTickets.length === 0 && (
+                <div className="flex justify-center items-center h-full bg-card border rounded-xl text-xl">
+                    No available tickets
+                </div>
+            )}
+        </div>
+    );
+};
+
+export { TicketTable };
