+++ src/data/tickets.sample.json
@@ -1,0 +1,325 @@
+[
+  {
+    "id": "TCKT-001",
+    "title": "Cannot access email",
+    "status": "Open",
+    "created_at": "2024-10-09",
+    "updated_at": "2024-10-09",
+    "assigned_to": "@bentot"
+  },
+  {
+    "id": "TCKT-002",
+    "title": "Printer not working",
+    "status": "Resolved",
+    "created_at": "2024-01-09",
+    "updated_at": "2024-01-10",
+    "assigned_to": "@bentot"
+  },
+  {
+    "id": "TCKT-003",
+    "title": "Request for new mouse",
+    "status": "Open",
+    "created_at": "2024-09-15",
+    "updated_at": "2024-09-16",
+    "assigned_to": "@bentot"
+  },
+  {
+    "id": "TCKT-004",
+    "title": "Monitor not displaying",
+    "status": "In progress",
+    "created_at": "2024-03-12",
+    "updated_at": "2024-03-13",
+    "assigned_to": "@bentot"
+  },
+  {
+    "id": "TCKT-005",
+    "title": "Request for external hard drive",
+    "status": "Closed",
+    "created_at": "2024-04-22",
+    "updated_at": "2024-04-23",
+    "assigned_to": "@bentot"
+  },
+  {
+    "id": "TCKT-006",
+    "title": "Projector not working",
+    "status": "Open",
+    "created_at": "2024-05-05",
+    "updated_at": "2024-05-05",
+    "assigned_to": "@bentot"
+  },
+  {
+    "id": "TCKT-007",
+    "title": "Request for additional storage",
+    "status": "Closed",
+    "created_at": "2024-06-21",
+    "updated_at": "2024-06-22",
+    "assigned_to": "@bentot"
+  },
+  {
+    "id": "TCKT-008",
+    "title": "Laptop battery issue",
+    "status": "Resolved",
+    "created_at": "2024-07-10",
+    "updated_at": "2024-07-11",
+    "assigned_to": "@bentot"
+  },
+  {
+    "id": "TCKT-009",
+    "title": "Request for new keyboard",
+    "status": "Open",
+    "created_at": "2024-08-18",
+    "updated_at": "2024-08-19",
+    "assigned_to": "@bentot"
+  },
+  {
+    "id": "TCKT-010",
+    "title": "Desktop not booting",
+    "status": "In progress",
+    "created_at": "2024-09-01",
+    "updated_at": "2024-09-02",
+    "assigned_to": "@bentot"
+  },
+
+  {
+    "id": "TCKT-011",
+    "title": "Software installation request",
+    "status": "In progress",
+    "created_at": "2024-02-09",
+    "updated_at": "2024-02-11",
+    "assigned_to": "@juantot"
+  },
+  {
+    "id": "TCKT-012",
+    "title": "Request for software upgrade",
+    "status": "Open",
+    "created_at": "2024-10-05",
+    "updated_at": "2024-10-06",
+    "assigned_to": "@juantot"
+  },
+  {
+    "id": "TCKT-013",
+    "title": "Email sync issue",
+    "status": "Open",
+    "created_at": "2024-10-14",
+    "updated_at": "2024-10-15",
+    "assigned_to": "@juantot"
+  },
+  {
+    "id": "TCKT-014",
+    "title": "Request for new laptop",
+    "status": "Open",
+    "created_at": "2024-10-16",
+    "updated_at": "2024-10-16",
+    "assigned_to": "@juantot"
+  },
+  {
+    "id": "TCKT-015",
+    "title": "Network configuration change",
+    "status": "Resolved",
+    "created_at": "2024-03-20",
+    "updated_at": "2024-03-21",
+    "assigned_to": "@juantot"
+  },
+  {
+    "id": "TCKT-016",
+    "title": "VPN setup",
+    "status": "Closed",
+    "created_at": "2024-04-15",
+    "updated_at": "2024-04-16",
+    "assigned_to": "@juantot"
+  },
+  {
+    "id": "TCKT-017",
+    "title": "Request for admin access",
+    "status": "Open",
+    "created_at": "2024-05-25",
+    "updated_at": "2024-05-26",
+    "assigned_to": "@juantot"
+  },
+  {
+    "id": "TCKT-018",
+    "title": "Printer setup",
+    "status": "Resolved",
+    "created_at": "2024-06-03",
+    "updated_at": "2024-06-04",
+    "assigned_to": "@juantot"
+  },
+  {
+    "id": "TCKT-019",
+    "title": "Monitor calibration",
+    "status": "In progress",
+    "created_at": "2024-07-12",
+    "updated_at": "2024-07-13",
+    "assigned_to": "@juantot"
+  },
+  {
+    "id": "TCKT-020",
+    "title": "System update failure",
+    "status": "Open",
+    "created_at": "2024-08-22",
+    "updated_at": "2024-08-22",
+    "assigned_to": "@juantot"
+  },
+
+  {
+    "id": "TCKT-021",
+    "title": "Password reset",
+    "status": "Resolved",
+    "created_at": "2024-04-09",
+    "updated_at": "2024-04-09",
+    "assigned_to": "@gwentot"
+  },
+  {
+    "id": "TCKT-022",
+    "title": "Monitor flickering",
+    "status": "Closed",
+    "created_at": "2024-05-09",
+    "updated_at": "2024-05-10",
+    "assigned_to": "@gwentot"
+  },
+  {
+    "id": "TCKT-023",
+    "title": "System crash during startup",
+    "status": "In progress",
+    "created_at": "2024-10-07",
+    "updated_at": "2024-10-08",
+    "assigned_to": "@gwentot"
+  },
+  {
+    "id": "TCKT-024",
+    "title": "Antivirus expired",
+    "status": "Resolved",
+    "created_at": "2024-10-17",
+    "updated_at": "2024-10-18",
+    "assigned_to": "@gwentot"
+  },
+  {
+    "id": "TCKT-025",
+    "title": "Slow internet connection",
+    "status": "Open",
+    "created_at": "2024-03-30",
+    "updated_at": "2024-03-31",
+    "assigned_to": "@gwentot"
+  },
+  {
+    "id": "TCKT-026",
+    "title": "Request for new monitor",
+    "status": "Closed",
+    "created_at": "2024-04-18",
+    "updated_at": "2024-04-19",
+    "assigned_to": "@gwentot"
+  },
+  {
+    "id": "TCKT-027",
+    "title": "Keyboard replacement",
+    "status": "Resolved",
+    "created_at": "2024-05-28",
+    "updated_at": "2024-05-29",
+    "assigned_to": "@gwentot"
+  },
+  {
+    "id": "TCKT-028",
+    "title": "Software license renewal",
+    "status": "Open",
+    "created_at": "2024-06-14",
+    "updated_at": "2024-06-15",
+    "assigned_to": "@gwentot"
+  },
+  {
+    "id": "TCKT-029",
+    "title": "Desktop overheating",
+    "status": "In progress",
+    "created_at": "2024-07-20",
+    "updated_at": "2024-07-21",
+    "assigned_to": "@gwentot"
+  },
+  {
+    "id": "TCKT-030",
+    "title": "Request for webcam",
+    "status": "Open",
+    "created_at": "2024-08-29",
+    "updated_at": "2024-08-29",
+    "assigned_to": "@gwentot"
+  },
+
+  {
+    "id": "TCKT-031",
+    "title": "VPN access request",
+    "status": "Closed",
+    "created_at": "2024-06-09",
+    "updated_at": "2024-06-10",
+    "assigned_to": "@kwintot"
+  },
+  {
+    "id": "TCKT-032",
+    "title": "Slow computer performance",
+    "status": "In progress",
+    "created_at": "2024-07-09",
+    "updated_at": "2024-07-12",
+    "assigned_to": "@kwintot"
+  },
+  {
+    "id": "TCKT-033",
+    "title": "Unable to connect to WiFi",
+    "status": "Resolved",
+    "created_at": "2024-10-10",
+    "updated_at": "2024-10-11",
+    "assigned_to": "@kwintot"
+  },
+  {
+    "id": "TCKT-034",
+    "title": "Projector not displaying",
+    "status": "In progress",
+    "created_at": "2024-10-19",
+    "updated_at": "2024-10-20",
+    "assigned_to": "@kwintot"
+  },
+  {
+    "id": "TCKT-035",
+    "title": "Request for new headset",
+    "status": "Open",
+    "created_at": "2024-03-25",
+    "updated_at": "2024-03-26",
+    "assigned_to": "@kwintot"
+  },
+  {
+    "id": "TCKT-036",
+    "title": "Network connectivity issue",
+    "status": "Open",
+    "created_at": "2024-04-27",
+    "updated_at": "2024-04-28",
+    "assigned_to": "@kwintot"
+  },
+  {
+    "id": "TCKT-037",
+    "title": "Request for new phone",
+    "status": "Closed",
+    "created_at": "2024-05-12",
+    "updated_at": "2024-05-13",
+    "assigned_to": "@kwintot"
+  },
+  {
+    "id": "TCKT-038",
+    "title": "System backup failed",
+    "status": "Resolved",
+    "created_at": "2024-06-30",
+    "updated_at": "2024-07-01",
+    "assigned_to": "@kwintot"
+  },
+  {
+    "id": "TCKT-039",
+    "title": "Request for new chair",
+    "status": "Open",
+    "created_at": "2024-08-05",
+    "updated_at": "2024-08-05",
+    "assigned_to": "@kwintot"
+  },
+  {
+    "id": "TCKT-040",
+    "title": "Mobile device setup",
+    "status": "Open",
+    "created_at": "2024-09-10",
+    "updated_at": "2024-09-11",
+    "assigned_to": "@kwintot"
+  }
+]
