+++ src/types.ts
@@ -1,14 +1,14 @@
 type TicketType = {
-  id: string;
-  title: string;
-  status: "Unassigned" | "In progress" | "Resolved" | "Closed";
-  created_at: string;
-  updated_at: string;
-  assigned_to: string;
+    id: string;
+    title: string;
+    status: "Open" | "In progress" | "Resolved" | "Closed";
+    created_at: string;
+    updated_at: string;
+    assigned_to: string;
 };
 
 type StatusSummaryType = {
-  status: string;
-  count: number;
+    status: "Open" | "In progress" | "Resolved" | "Closed";
+    count: number;
 };
 export { type TicketType, type StatusSummaryType };
